extract Jwt and Urn classes from core to external files
