+++ lib/core/lib/ros/core.rb
@@ -23,6 +23,8 @@ require_relative 'url_builder'
 require_relative 'jsonapi_authorization/authorizer'
 require_relative 'api_token_strategy'
 require_relative 'routes'
+require_relative 'urn'
+require_relative 'jwt'
 require_relative '../migrations'
 
 require 'ros/core/engine'
@@ -65,92 +67,6 @@ module Ros
 
     # By default all services exclude only the Tenant model from schemas
     def excluded_models; %w[Tenant] end
-  end
-
-  # TODO: Authorize method
-  # TODO: scope value is the subject's policies; What if the subject's policies change after token issued?
-  # It could be that every token expires in 10 minutes or something which means that the auth strategy
-  # would check and then re-issue a token as long as the user is still valid;
-  # this would update permissions at that point
-  # or it would need to check if the user has been updated since the token was issued
-  # def self.issue(iss: Ros::Sdk.service_endpoints['iam'], sub:, scope:)
-  #   issued_at = Time.now.to_i
-  #   token = { iss: iss, aud: ['this_domain'], sub: sub, scope: scope, iat: issued_at }
-  #   token.merge!(exp: issued_at + expires_in) if expires_in = Settings.dig(:jwt, :token_expires_in_seconds)
-  #   token
-  # end
-  class Jwt
-    attr_reader :claims, :token
-
-    def initialize(payload)
-      if payload.is_a? Hash # From IAM::User, IAM::Root
-        @claims = payload.merge(default_payload)
-      else # From a bearer token
-        @token = payload.gsub('Bearer ', '')
-        decode
-      end
-    end
-
-    def default_payload
-      issued_at = Time.now.to_i
-      token = (expires_in = Settings.dig(:jwt, :token_expires_in_seconds)) ? { exp: issued_at + expires_in } : {}
-      token.merge(iss: iss, aud: aud, iat: issued_at)
-    end
-
-    def add_claims(claims = {})
-      @claims.merge!(claims.select{ |k, v| k.to_s.in?(valid_claims) })
-      self
-    end
-
-    def valid_claims; @valid_claims ||= (Settings.dig(:jwt, :valid_claims) || []) end
-
-    # TODO: Set audience from the issuer's domain name
-    def aud; Settings.jwt.aud end
-    def iss; Settings.jwt.iss end
-
-    def encode
-      @token = JWT.encode(claims, encryption_key, alg)
-    end
-
-    def decode
-      @claims = JWT.decode(token, encryption_key, alg).first
-    end
-
-    def encryption_key; Settings.jwt.encryption_key end
-
-    def alg; Settings.jwt.alg end
-  end
-
-  Urn = Struct.new(:txt, :partition_name, :service_name, :region, :account_id, :resource) do
-    def self.from_urn(urn_string)
-      urn_array = urn_string.split(':')
-      new(*urn_array)
-    end
-
-    def self.from_jwt(token)
-      jwt = Jwt.new(token)
-      return unless (urn_string = jwt.decode['sub'])
-
-      from_urn(urn_string)
-    # NOTE: Intentionally swallow decode error and return nil
-    # rubocop:disable Lint/HandleExceptions
-    rescue JWT::DecodeError
-    end
-    # rubocop:enable Lint/HandleExceptions
-
-    def resource_type; resource.split('/').first end
-
-    def resource_id; resource.split('/').last end
-
-    def model_name; resource_type.classify end
-
-    def model; model_name.constantize end
-
-    # rubocop:disable Rails/DynamicFindBy
-    def instance; model.find_by_urn(resource_id) end
-    # rubocop:enable Rails/DynamicFindBy
-
-    def to_s; to_a.join(':') end
   end
 
   # Failure response to return JSONAPI error message when authentication fails
